+++ test/test_util.py
@@ -108,6 +108,130 @@ def test_split_keywords(kwstr, expected):
             "flit.wheel",
         ],
     ),
+
+    (
+        [
+            'cmarkgfm-0.4.2.dist-info/WHEEL',
+            'cmarkgfm-0.4.2.dist-info/top_level.txt',
+            'cmarkgfm-0.4.2.dist-info/METADATA',
+            'cmarkgfm-0.4.2.dist-info/RECORD',
+            'cmarkgfm/_cmark.abi3.so',
+            'cmarkgfm/cmark_module.h',
+            'cmarkgfm/__init__.py',
+            'cmarkgfm/cmark.py',
+            'cmarkgfm/build_cmark.py',
+            'cmarkgfm/cmark.cffi.h',
+        ],
+        [
+            'cmarkgfm',
+            'cmarkgfm._cmark',
+            'cmarkgfm.build_cmark',
+            'cmarkgfm.cmark',
+        ],
+    ),
+
+    (
+        [
+            'cmarkgfm-0.4.2.dist-info/WHEEL',
+            'cmarkgfm-0.4.2.dist-info/top_level.txt',
+            'cmarkgfm-0.4.2.dist-info/METADATA',
+            'cmarkgfm-0.4.2.dist-info/RECORD',
+            'cmarkgfm/_cmark.cp37-win_amd64.pyd',
+            'cmarkgfm/cmark_module.h',
+            'cmarkgfm/__init__.py',
+            'cmarkgfm/cmark.py',
+            'cmarkgfm/build_cmark.py',
+            'cmarkgfm/cmark.cffi.h',
+        ],
+        [
+            'cmarkgfm',
+            'cmarkgfm._cmark',
+            'cmarkgfm.build_cmark',
+            'cmarkgfm.cmark',
+        ],
+    ),
+
+    (
+        [
+            'cmarkgfm-0.4.2.dist-info/WHEEL',
+            'cmarkgfm-0.4.2.dist-info/top_level.txt',
+            'cmarkgfm-0.4.2.dist-info/METADATA',
+            'cmarkgfm-0.4.2.dist-info/RECORD',
+            'cmarkgfm/_cmark.so',
+            'cmarkgfm/cmark_module.h',
+            'cmarkgfm/__init__.py',
+            'cmarkgfm/cmark.py',
+            'cmarkgfm/build_cmark.py',
+            'cmarkgfm/cmark.cffi.h',
+        ],
+        [
+            'cmarkgfm',
+            'cmarkgfm._cmark',
+            'cmarkgfm.build_cmark',
+            'cmarkgfm.cmark',
+        ],
+    ),
+
+    (
+        [
+            'mxnet_coreml_converter-0.1.0a7.data/purelib/converter/__init__.py',
+            'mxnet_coreml_converter-0.1.0a7.data/purelib/converter/_add_pooling.py',
+            'mxnet_coreml_converter-0.1.0a7.data/purelib/converter/_layers.py',
+            'mxnet_coreml_converter-0.1.0a7.data/purelib/converter/_mxnet_converter.py',
+            'mxnet_coreml_converter-0.1.0a7.data/purelib/converter/utils.py',
+            'mxnet_coreml_converter-0.1.0a7.data/scripts/mxnet_coreml_converter.py',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/DESCRIPTION.rst',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/metadata.json',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/top_level.txt',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/WHEEL',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/METADATA',
+            'mxnet_coreml_converter-0.1.0a7.dist-info/RECORD',
+        ],
+        [
+            'converter',
+            'converter._add_pooling',
+            'converter._layers',
+            'converter._mxnet_converter',
+            'converter.utils',
+        ],
+    ),
+
+    (
+        [
+            'foo-1.0.data/platlib/foo/__init__.py',
+            'foo-1.0.data/platlib/foo/def.py',
+            'foo-1.0.data/platlib/foo/has-hyphen.py',
+            'foo-1.0.data/platlib/foo/extra.ext.py',
+            'foo-1.0.dist-info/METADATA',
+            'foo-1.0.dist-info/WHEEL',
+            'foo-1.0.dist-info/RECORD',
+            'foo-1.0.dist-info/glarch.py',
+        ],
+        ['foo'],
+    ),
+
+    (
+        [
+            'Acquisition/Acquisition.h',
+            'Acquisition/_Acquisition.c',
+            'Acquisition/_Acquisition.pyd',
+            'Acquisition/__init__.py',
+            'Acquisition/interfaces.py',
+            'Acquisition/tests.py',
+            'Acquisition-4.6.dist-info/LICENSE.txt',
+            'Acquisition-4.6.dist-info/METADATA',
+            'Acquisition-4.6.dist-info/WHEEL',
+            'Acquisition-4.6.dist-info/top_level.txt',
+            'Acquisition-4.6.dist-info/RECORD',
+        ],
+        [
+            'Acquisition',
+            'Acquisition._Acquisition',
+            'Acquisition.interfaces',
+            'Acquisition.tests',
+        ],
+    ),
+
 ])
 def test_extract_modules(filelist, modules):
     assert extract_modules(filelist) == modules